requirements.txt
dateparser==0.7.0
dj-database-url==0.5.0
Django==2.0.2
django-appconf==1.0.2
django-background-tasks==1.1.13
django-compat==1.0.15
django-measurement==3.0.0
django-money==0.12.3
measurement==2.0.1
mpmath==1.0.0
psycopg2==2.7.4
py-moneyed==0.7.0
python-dateutil==2.7.0
pytz==2018.3
regex==2018.2.21
selenium==3.11.0
six==1.11.0
sympy==1.1.1
tzlocal==1.5.1
